Disable no-bitwise lint rule; Add lint ignore
According to eslint [docs], no-bitwise rule is set to avoid confusion of
bitwise operators and logical operators, and under the assumption that
bitwise is rarely used. Considering the importance in this plugin, I
have to disable this rule for readability, in cost of consistence with
markdown-it lint rules.

There's no intention to have linting on test/, so it's added to .eslintignore

[docs]: https://eslint.org/docs/rules/no-bitwise